Cleaned output files.

+++ src/main/scala/multipleStrains/models/ClassRoom.scala
@@ -3,7 +3,7 @@ package multipleStrains.models
 import com.bharatsim.engine.models.Network
 
 
-case class ClassRoom(classRoomId: Int) extends Network {
+case class ClassRoom(classRoomId: Long) extends Network {
   override def getContactProbability(): Double = 1
 
   addRelation[Person]("CLASS_TEACHES")
